bench: 9292266

+++ src/structs.h
@@ -50,7 +50,7 @@ typedef struct {
 // transposition table data structure
 typedef struct {
     U64 hashKey;     // "almost" unique chess position identifier
-    int8_t depth;       // current search depth
+    int16_t depth;       // current search depth
     uint8_t flag;       // flag the type of node (fail-high(score >= beta)/fail-low(score < alpha))
     int16_t score;       // score (alpha/beta/PV)
     int bestMove;
